+++ changelog.md
@@ -1,5 +1,6 @@
-#### 0.0.5 (2021-09-25)
+#### 0.0.5 (2021-09-27)
 
+- don't convert url if already requested language (#8e46ee4)
 - Refactor URL comparisons 💖 (#c7fc073)
 - Merged in fix-check-post-for-null (pull request #7)  check for null value before using $post object  Approved-by: Rasso Hilber (#f5b0828)
 - New Filter `format_acf_field_link` + New Function to detect links to physical locations 🧐 (#b8da2cb)